+++ src/com/power/oj/user/UserService.java
@@ -125,7 +125,7 @@ public class UserService
     newUser.set("phone", HtmlEncoder.text(userModel.getStr("phone")));
     newUser.set("gender", HtmlEncoder.text(userModel.getStr("gender")));
     newUser.set("language", userModel.getInt("language"));
-    newUser.set("qq", userModel.getInt("qq"));
+    newUser.set("qq", userModel.getLong("qq"));
     newUser.set("mtime", OjConfig.timeStamp);
     
     return newUser.update();
